8/15 Update #1
Game can now detect how many eggs are sorted correctly and incorrectly in each incubator. Also, adjusted resultsScript.cs so the game knows when all the eggs in the level have been put into an incubator, not just a fixed number of eggs that have been put into an incubator.